refactor: decoupling datapane
