+++ superset-frontend/src/explore/components/DataTablesPane/index.ts
@@ -16,12 +16,5 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { useSelector } from 'react-redux';
-import { ExplorePageState } from '../reducers/getInitialState';
-
-export const useOriginalFormattedTimeColumns = (datasourceId?: string) =>
-  useSelector<ExplorePageState, string[]>(state =>
-    datasourceId
-      ? state?.explore?.originalFormattedTimeColumns?.[datasourceId] ?? []
-      : [],
-  );
+export { DataTablesPane } from './DataTablesPane';
+export { DataTable } from './DataTable';
