+++ superset-frontend/src/explore/components/DataTablesPane/types.ts
@@ -1,0 +1,65 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import {
+  Datasource,
+  GenericDataType,
+  JsonObject,
+  QueryFormData,
+} from '@superset-ui/core';
+import { ExploreActions } from 'src/explore/actions/exploreActions';
+
+export interface DataTablesPaneProps {
+  queryFormData: QueryFormData;
+  datasource: Datasource;
+  queryForce: boolean;
+  ownState?: JsonObject;
+  onCollapseChange: (isOpen: boolean) => void;
+  errorMessage?: JSX.Element;
+  actions: ExploreActions;
+}
+
+export interface DataTableProps {
+  isRequest: boolean;
+  resultType: 'results' | 'samples';
+  queryFormData: QueryFormData;
+  datasource?: Datasource;
+  queryForce: boolean;
+  ownState?: JsonObject;
+  errorMessage?: React.ReactElement;
+  actions?: ExploreActions;
+  dataSize?: number;
+}
+
+export interface responseData {
+  data: Record<string, any>[][];
+  colnames: string[];
+  coltypes?: GenericDataType[];
+  isLoading: boolean;
+  responseError: string;
+}
+
+export interface requestPayload {
+  isRequest: boolean;
+  resultType: 'results' | 'samples';
+  queryFormData: QueryFormData;
+  datasource?: Datasource;
+  queryForce: boolean;
+  ownState?: JsonObject;
+  actions?: ExploreActions;
+}
